Merge monorepo package
This Dart package was originally created in the enspyrco monorepo -
apparently this separate repo was created and the code was copied over,
then the two repos diverged with new work being done in the monorepo.

This commit merges the changes from the monorepo package into this
separate repo and the package was removed from the monorepo in
https://github.com/enspyrco/monorepo/pull/345

This repo is now the source of truth for the discord_api_client package.